+++ fractals/src/main/java/me/catmousedog/fractals/fractals/types/iterative/TestFractal.java
@@ -1,5 +1,7 @@
 package me.catmousedog.fractals.fractals.types.iterative;
 
+import java.util.Properties;
+
 import org.jetbrains.annotations.NotNull;
 
 import me.catmousedog.fractals.fractals.Fractal;