+++ fractals/src/main/java/me/catmousedog/fractals/fractals/types/normalized/NormalizedJulia.java
@@ -1,4 +1,4 @@
-package me.catmousedog.fractals.fractals.types.iterative;
+package me.catmousedog.fractals.fractals.types.normalized;
 
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseMotionListener;
@@ -6,8 +6,7 @@ import java.util.Properties;
 
 import me.catmousedog.fractals.fractals.Fractal;
 import me.catmousedog.fractals.fractals.filters.Filter;
-import me.catmousedog.fractals.fractals.filters.IterativeLinearFilter;
-import me.catmousedog.fractals.fractals.filters.IterativePeriodicFilter;
+import me.catmousedog.fractals.fractals.filters.LogPeriodicFilter;
 import me.catmousedog.fractals.main.Settings;
 import me.catmousedog.fractals.ui.components.Item;
 import me.catmousedog.fractals.ui.components.concrete.CheckBox;
@@ -17,18 +16,20 @@ import me.catmousedog.fractals.ui.components.concrete.TextFieldDouble;
 /**
  * Number = Integer
  */
-public final class IterativeJulia extends Fractal {
+public final class NormalizedJulia extends Fractal {
 
+	private int offset;
 	private double jx, jy;
 
-	public IterativeJulia(Settings settings) {
+	public NormalizedJulia(Settings settings) {
 		super(settings);
 	}
 
-	private IterativeJulia(Settings settings, Fractal fractal, double jx, double jy) {
+	private NormalizedJulia(Settings settings, Fractal fractal, double jx, double jy, int offset) {
 		super(settings, fractal);
 		this.jx = jx;
 		this.jy = jy;
+		this.offset = offset;
 	}
 
 	@Override
@@ -36,6 +37,7 @@ public final class IterativeJulia extends Fractal {
 		double x = cx, y = cy;
 		double tx;
 		double t1, t2;
+		double p;
 
 		for (int i = 0; i < iterations; i++) {
 			tx = x;
@@ -43,12 +45,14 @@ public final class IterativeJulia extends Fractal {
 			t1 = x * x;
 			t2 = y * y;
 
-			if (t1 + t2 > bailout)
-				return 255 * (iterations - i) / iterations;
+			p = Math.log(t1 + t2) / 2;
+
+			if (p > bailout) {
+				return i + offset - Math.log(p) / Math.log(2);
+			}
 
 			x = t1 - t2 + jx;
 			y = 2 * tx * y + jy;
-
 		}
 		return 0;
 	}
@@ -69,18 +73,18 @@ public final class IterativeJulia extends Fractal {
 
 	@Override
 	public String informalName() {
-		return "Iterative Julia Set";
+		return "Normalized Julia Set";
 	}
 
 	@Override
 	public String fileName() {
-		return "IterativeJulia";
+		return "NormalizedJulia";
 	}
 
 	@Override
 	public String getTip() {
-		return "<html>The second order julia set (<i>z²+c<i/>) generated using an escape time algorithm."
-				+ "<br>This allows for deep zooms but creates aliasing effects, generally has the shortest generating time."
+		return "<html>The second order julia set (<i>z²+c<i/>) generated using an normalized iteration counts."
+				+ "<br>This allows for deep zooms but is slower."
 				+ "<br>Dragging the mouse will change the fixed point used to generate this set.</html>";
 	}
 
@@ -92,7 +96,7 @@ public final class IterativeJulia extends Fractal {
 
 	@Override
 	protected void initFractal() {
-		filters = new Filter[] { new IterativeLinearFilter(this), new IterativePeriodicFilter(this) };
+		filters = new Filter[] { new LogPeriodicFilter(this) };
 		filter = filters[0];
 		mouse = new MouseMotionListener() {
 			@Override
@@ -114,9 +118,9 @@ public final class IterativeJulia extends Fractal {
 
 		Padding p5 = new Padding(5);
 
-		mousejcx = new CheckBox.Builder("mouse listener").setTip(
-				"<html>If enabled, the user can click and drag the mouse along the canvas<br>"
-				+ " to change the fixed julia point.</html>")
+		mousejcx = new CheckBox.Builder("mouse listener")
+				.setTip("<html>If enabled, the user can click and drag the mouse along the canvas<br>"
+						+ " to change the fixed julia point.</html>")
 				.build();
 
 		jxjtf = new TextFieldDouble.Builder().setLabel("Jx").setTip("The fixed julia point's x-coordinate")
@@ -128,7 +132,13 @@ public final class IterativeJulia extends Fractal {
 	}
 
 	@Override
+	public void setProperties(Properties properties) {
+		super.setProperties(properties);
+		offset = Integer.parseInt(properties.getProperty("offset"));
+	}
+
+	@Override
 	public Fractal clone() {
-		return new IterativeJulia(settings, this, jx, jy);
+		return new NormalizedJulia(settings, this, jx, jy, offset);
 	}
 }